Updated to support EventServer conversion to Android Service
init.rc now creates the sensors socket under the net_admin user so it can be
accessed by an app with the NET_ADMIN system permission.
init.rc now does not attempt to start the EventServer, which is now an Android Service that starts
on boot.
LocationHelper now starts on SVMP boot message, not system boot message.

Change-Id: I66076cf3a729bebbf705b4ca53c1268f75f905ba