Split build script to workaround repo issue
The repo tool's copy directive that is used to copy the build script
from device/mitre/svmp into the AOSP root does not seem to always pick
up changes to the build script and copy the new version into root. Fix
by making build-svmp.sh a wrapper that calls the real script directly
from the device tree directory so it's always in sync. The wrapper will
never need to change.

Change-Id: Id4a381bb613fc26be1d16f9dd7e42b0cc71bf03b