Fix the intermittent surfaceflinger bootloop
Update init.rc to fix race condition with mediaserver and
surfaceflinger.  Mediaserver fix moves the startup class to svmp to
ensure that it doesn't start too soon. For the surfaceflinger issue,
for some reason surfaceflinger somehow reads incorrect values from the
fbdriver. If this happens, it will crash. We add a rule to restart
fbset when this occurs and this seems to resolve itself.

Change-Id: I0c3dd7d7ab23b73761e311b37749e7d91f0cec50