Initial commit of "fake" microphone device.
This device is used to "pickup" audio generated from applications
running within the VM. The scheme works as follows: An audio
application like an mp3 player will playback to the virtual speaker.
The virtual speaker will in turn write to a shared memory in the
form of an mmaped file. The microphone will read from the same
shared memory.

Change-Id: Ifda5e7ee19315946e998fc285a1ec11cdeffec57