Refactored SVMP VM image build rules
Moved all of the VM disk image building rules out of diskinstaller
and into image_build/svmp-images.mk here in the device tree.

New make targets:
  svmp_system_disk svmp_system_disk_[vdi|vmdk|qcow2]
  svmp_data_disk svmp_data_disk_[vdi|vmdk|qcow2]

All images are built using editdisklbl disk layout configurations. This
includes the user data disk, which is no longer a manually pre-configured
image.

Moved the cache partition onto the system disk from the data disk. The
data stored in /cache is ephemeral and not portable across ROM versions
so we probably shouldn't be putting it in the persistent data image. The
data image format is left intact the way it was before with a cache
partition present in order to retain backwards compatibility with old
existing data partitions. Worst case, we waste a GB in one image or
the other.

Change-Id: I768c1d3f8d886c4ce390fe2aef2249c5fdc8e509