Add grub menu.lst to the ramdisk image
Xen-based systems booting in paravirt mode require a grub menu.lst file
be present in the image for pvgrub to read.

Non-xen images will still use the embedded grub. Note, however, that the
internal grub does NOT use the menu.lst but instead uses a hard coded
string in external/grub/stage2/preset_menu.c. For it to use a menu.lst
file in the filesystem we would need to add a stage1_5 image for ext2,
which is left as an exercise for the reader.

Change-Id: I1cad489118c1757b94a72d50915c7e5644b0df17