Display VM IP info on console
The serial console is often a pain to get at (if not impossible to) on
certain virtualization platforms. This patch prints the network iface
info to the graphical console (/dev/tty0) and switches off the kernel's
automatic console blanking (since there's no way to wake the console
back up once blanking activates).

The full dmesg log and interactive shell is left on the first serial
port, /dev/ttyS0.

Change-Id: I7e564297b58fc02ac4fd948c40300decd98772af